[front] chore: move PostHog tracking to API layer
